face-recognition/camera_ocean/requirements.txt
opencv-python
imutils
click
requests
django-environ
raven